Move RenderingTypePredictor to playwright utils
