+++ packages/playwright-crawler/src/index.ts
@@ -6,3 +6,4 @@ export * from './internals/adaptive-playwright-crawler';
 export * as playwrightUtils from './internals/utils/playwright-utils';
 export * as playwrightClickElements from './internals/enqueue-links/click-elements';
 export type { DirectNavigationOptions as PlaywrightDirectNavigationOptions } from './internals/utils/playwright-utils';
+export type { RenderingType } from './internals/utils/rendering-type-prediction';
